optimize corp afordable upgrade level calculation
